Seeders, Factories

+++ database/migrations/2023_02_14_142450_create_products_table.php
@@ -19,8 +19,8 @@ return new class extends Migration
             $table -> string('code', 5) -> unique();
             $table -> string('name', 64);
             $table -> text('description') -> nullable();
-            $table -> integer('price') -> unsigned();
-            $table -> integer('weight') -> unsigned() -> default(0);
+            $table -> float('price') -> unsigned();
+            $table -> float('weight') -> unsigned() -> default(0);
             
             $table->timestamps();
         });
